調前腳與腳側

+++ apps/addonsExamples/networkTcpServerExampleMarionette/src/testApp.h
@@ -135,7 +135,7 @@ class testApp : public ofSimpleApp
 		ofSoundPlayer  vocals;
 		#endif
 		float 			synthPosition;
-		long			totalSec;
+		double			totalSec;
 
 		ofTrueTypeFont  franklinBook;
